+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 RC
 --
 
+
 + jinja2 float casts are no longer just removed, but now 
   really cast a number to a float.
+
++ remove newline at the end of a rendered template
 
 
 v1.0.0